refactor: first batch moving dream syntax to post
- also splits db reads and writes into domains
