+++ lib/db/migrations.js
@@ -1,15 +1,15 @@
 import { ObjectID } from "bson";
 import { getInsights } from "../insights";
-import { getDreamsCollection } from "../mongodb";
-import { decryptDream } from "../transformations";
-import { hasCommentedOnDream } from "./reads";
+import { getPostsCollection } from "../mongodb";
+import { decryptPost } from "../transformations";
+import { hasCommentedOnPost } from "./reads";
 import { generateCompletion } from "./writes";
 
 /**
  * @dataCorrection
  */
 export async function addMissingWordFreqData() {
-  const collection = await getDreamsCollection();
+  const collection = await getPostsCollection();
 
   const cursor = collection.find();
 
@@ -24,7 +24,7 @@ export async function addMissingWordFreqData() {
 
   await Promise.all([
     ...privateDeams.map((data) => {
-      const decrypted = decryptDream(data.dream);
+      const decrypted = decryptPost(data.dream);
       const insights = getInsights(decrypted.text);
 
       collection.updateOne(
@@ -58,7 +58,7 @@ export async function addMissingWordFreqData() {
  * @dataCorrection
  */
 export async function addMissingAICommentData() {
-  const collection = await getDreamsCollection();
+  const collection = await getPostsCollection();
 
   const cursor = collection.find();
 
@@ -74,10 +74,10 @@ export async function addMissingAICommentData() {
   await Promise.all([
     ...privateDeams.map(async (data) => {
       ["Sonio", "Sonia"].map(async (aiName) => {
-        const hasCommented = await hasCommentedOnDream(aiName, data._id);
+        const hasCommented = await hasCommentedOnPost(aiName, data._id);
 
         if (!hasCommented) {
-          const decrypted = decryptDream(data.dream);
+          const decrypted = decryptPost(data.dream);
           return generateCompletion(
             data.dream._id,
             decrypted.text,
@@ -89,7 +89,7 @@ export async function addMissingAICommentData() {
     }),
     ...publicDreams.map(async (data) => {
       ["Sonio", "Sonia"].map(async (aiName) => {
-        const hasCommented = await hasCommentedOnDream(aiName, data._id);
+        const hasCommented = await hasCommentedOnPost(aiName, data._id);
 
         if (!hasCommented) {
           return generateCompletion(
