+++ lib/mongodb.js
@@ -73,9 +73,9 @@ export async function getMongoClient() {
 }
 
 /**
- * Connects to the database and gets the dreams collection.
+ * Connects to the database and gets the posts collection.
  */
-export async function getDreamsCollection() {
+export async function getPostsCollection() {
   return (await connectToDatabase()).collection("dreams");
 }
