+++ lib/transformations.js
@@ -1,15 +1,15 @@
 import { decrypt, encrypt } from "./encryption";
 
-export function decryptDream(dream) {
+export function decryptPost(post) {
   return {
-    text: decrypt(dream.text),
-    html: decrypt(dream.html),
+    text: decrypt(post.text),
+    html: decrypt(post.html),
   };
 }
 
-export function encryptDream(dream) {
+export function encryptPost(post) {
   return {
-    text: encrypt(dream.text),
-    html: encrypt(dream.html),
+    text: encrypt(post.text),
+    html: encrypt(post.html),
   };
 }
