+++ lib/validations.js
@@ -1,3 +1,3 @@
-export async function isDreamOwner(dream, user) {
+export async function isPostOwner(dream, user) {
   return dream.userId === user._id;
 }
